clean up more around the place
