+++ mystbin/backend/routers/admin.py
@@ -25,8 +25,9 @@ import psutil
 from asyncpg import Record
 from fastapi import APIRouter
 from fastapi.responses import Response, UJSONResponse
-from fastapi_models import MystbinRequest
 from models import errors, responses
+
+from mystbin.backend.mystbin_models import MystbinRequest
 from utils.db import _recursive_hook as recursive_hook
 from utils.ratelimits import limit
