+++ mystbin/backend/routers/apps.py
@@ -24,8 +24,9 @@ from typing import Dict, Optional, Union
 import yarl
 from fastapi import APIRouter
 from fastapi.responses import Response, UJSONResponse
-from fastapi_models import MystbinRequest
 from models import responses
+
+from mystbin.backend.mystbin_models import MystbinRequest
 from utils.embed import Embed
 from utils.ratelimits import limit
