+++ mystbin/backend/routers/pastes.py
@@ -28,8 +28,9 @@ from typing import Dict, List, Optional, Union
 from asyncpg import Record
 from fastapi import APIRouter
 from fastapi.responses import UJSONResponse
-from fastapi_models import MystbinRequest
 from models import errors, payloads, responses
+
+from mystbin.backend.mystbin_models import MystbinRequest
 from utils.db import _recursive_hook as recursive_hook
 from utils.ratelimits import limit
